refactor: clarify difference between prettier plugin and config
