+++ packages/eslint-config-bases/src/bases/prettier-config.js
